payloads/SeaBIOS: apply bootsplash patch
Patch SeaBIOS to not start in console mode if a bootsplash
image is to be displayed, eliminating the flicker from
starting in conole mode then immediately switchint to the bootslpash.

Adjust application of patch file so as to make for a reproducible build.
The committer name, email and date needs to match the author's
otherwise on every build, seabios will have a different commit
hash for HEAD, which can be problematic for reproducable builds
but also could confuse people if every time they type 'make', they
get a different hash of their rom

Change-Id: I493c90e73722aa8046262fc0a97fcdeab12982a4
Signed-off-by: Matt DeVillier <matt.devillier@puri.sm>