Add cfgs, bootsplash for librem_mini
Librem Mini still uses a VBIOS for display init, needs a differently-
sized bootsplash image than other Librems using libgfxinit.

Add an extry for the NVMe drive on the LM to the default bootorder.

Add an 8s SeaBIOS boot menu delay for the LM to give the display enough
time to init, show the boot splash, and for the user to opt to show the menu.

UEFI and HEADS configs are for testing/convenience, not distribution.

Signed-off-by: Matt DeVillier <matt.devillier@puri.sm>
Change-Id: I30462cf05b365ee224ca44d5f90ae798f185473a